requirements.txt
Django==3.0.6
django-cors-headers==3.1.0
django-crontab==0.7.1
django-filter==2.2.0
django-redis==4.11.0
djangorestframework==3.10.2
eventregistry==8.6.1
marshmallow==3.6.0
psycopg2-binary==2.8.5
pytest==5.4.2
pytz==2019.2
redis==3.5.2
gunicorn==20.0.4
